Reverted to code to only one file, and changed the send value to total count
